dom/plugins/test/mochitest/test_instance_re-parent.html

<!DOCTYPE html>
<html>
<head>
  <title>Test re-parentinging an instance's DOM node</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="runTests()">
  <p id="display"></p>

  <div id="div1">
    <!-- This embed has to have a "src" attribute. Part of this test involves seeing if we
         properly restart plugins that have been added back to a document without a change
         in URL. Not re-loading an object when the URL hasn't changed is a shortcut used for
         some object types. Without a URL, this won't be tested. -->
    <embed id="plugin1" src="loremipsum.txt" type="application/x-test" width="200" height="200"></embed>
  </div>
  <div id="div2">
  </div>

  <script type="application/javascript;version=1.8">
  SimpleTest.waitForExplicitFinish();

  var exceptionThrown = false;
  var p = document.getElementById('plugin1');
  var d1 = document.getElementById('div1');
  var d2 = document.getElementById('div2');

  var destroyed = false;
  function onDestroy() {
    destroyed = true;
  }

  function testsFinished() {
    p.stopWatchingInstanceCount();
    SimpleTest.finish();
  }

  function continueTests2() {
    try {
      is(p.getInstanceCount(), 1, "One new instance should have been created at this point.");
    } catch (e) {
      exceptionThrown = true;
    }
    is(exceptionThrown, false, "Testing for exception getting instance count from plugin.");

    testsFinished();
  }

  function continueTests1() {
    // Adding to the document will kick off the plugin load. Need to have a timeout
    // before we can safely call into it.
    d2.appendChild(p);
    setTimeout('continueTests2();', 0);
  }

  function runTests() {
    // First tests involve moving the instance from one div to another.
    p.startWatchingInstanceCount();
    p.callOnDestroy(onDestroy);

    try {
      d1.removeChild(p);
    } catch (e) {
      exceptionThrown = true;
    }
    is(exceptionThrown, false, "Testing for exception after removeChild.");

    try {
      d2.appendChild(p);
    } catch (e) {
      exceptionThrown = true;
    }
    is(exceptionThrown, false, "Testing for exception after appendChild.");

    is(destroyed, false, "No instances should have been destroyed at this point.");
    is(p.getInstanceCount(), 0, "No new instances should have been created at this point.");

    // Now remove the instance from the document and let it die.
    d2.removeChild(p);
    setTimeout('continueTests1();', 0);
  }
  </script>
</body>
</html>
